refactor(network): convert to standard test module layout
