+++ zebra-network/src/config/tests/vectors.rs
@@ -1,0 +1,48 @@
+//! Fixed test vectors for zebra-network configuration.
+
+use static_assertions::const_assert;
+
+use crate::{
+    constants::{INBOUND_PEER_LIMIT_MULTIPLIER, OUTBOUND_PEER_LIMIT_MULTIPLIER},
+    Config,
+};
+
+#[test]
+fn parse_config_listen_addr() {
+    zebra_test::init();
+
+    let fixtures = vec![
+        ("listen_addr = '0.0.0.0'", "0.0.0.0:8233"),
+        ("listen_addr = '0.0.0.0:9999'", "0.0.0.0:9999"),
+        (
+            "listen_addr = '0.0.0.0'\nnetwork = 'Testnet'",
+            "0.0.0.0:18233",
+        ),
+        (
+            "listen_addr = '0.0.0.0:8233'\nnetwork = 'Testnet'",
+            "0.0.0.0:8233",
+        ),
+    ];
+
+    for (config, value) in fixtures {
+        let config: Config = toml::from_str(config).unwrap();
+        assert_eq!(config.listen_addr.to_string(), value);
+    }
+}
+
+/// Make sure the peer connection limits are consistent with each other.
+#[test]
+fn ensure_peer_connection_limits_consistent() {
+    zebra_test::init();
+
+    // Zebra should allow more inbound connections, to avoid connection exhaustion
+    const_assert!(INBOUND_PEER_LIMIT_MULTIPLIER > OUTBOUND_PEER_LIMIT_MULTIPLIER);
+
+    let config = Config::default();
+
+    assert!(
+        config.peerset_inbound_connection_limit() - config.peerset_outbound_connection_limit()
+            >= 50,
+        "default config should allow more inbound connections, to avoid connection exhaustion",
+    );
+}
